backend/aws-setup/requirements.txt
# Requirements for AWS Database Setup and Testing

# PostgreSQL
psycopg2-binary==2.9.9

# AWS SDK
boto3==1.34.14
botocore==1.34.14

# Environment variables
python-dotenv==1.0.0

# HTTP requests for API testing
requests==2.31.0

# Data privacy and encryption
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Data validation
pydantic==2.5.3

# Utilities
colorama==0.4.6  # For colored terminal output